release/requirements.txt
pillow>=10.0.0
numpy>=1.24.0
pyqt5>=5.15.0
